requirements.txt
streamlit~=1.50.0
pdfplumber~=0.11.0
pandas~=2.3.0
plotly~=6.6.0
gspread~=6.2.0
google-auth~=2.49.0
google-api-python-client~=2.100.0
fpdf2~=2.8.0
